Refactor inspection context management and remove legacy ContextService
- Deleted PartContext class as part of the transition to InspectionContext.
- Updated PartUtils and various operator rules to utilize InspectionContextHolder for context management.
- Refactored ApiGetRule, DaxQueryRule, DfsGetRule, and ScannerApiRule to reference InspectionContext instead of ContextService.
- Removed ContextServiceConcurrencyTests as it is no longer relevant.
- Modified OperatorProgressReportingTests to use instance variables for context instead of static ContextService properties.
- Introduced InspectionEngine to encapsulate per-run state and manage inspection execution.
- Added RulesFileLoader to handle deserialization of rules files without relying on static state.
